+++ src/background/utils/buildinProvider.ts
@@ -82,8 +82,9 @@ export class EthereumProvider extends EventEmitter {
     const mapMethod = underline2Camelcase(method);
     const currentAccount = preferenceService.getCurrentAccount()!;
     let networkId = CHAINS[CHAINS_ENUM.ETH].id.toString();
+    //todo
     if (currentAccount.type === KEYRING_CLASS.GNOSIS) {
-      networkId = wallet.getGnosisNetworkId(currentAccount.address);
+      // networkId = wallet.getGnosisNetworkId(currentAccount.address);
     } else {
       networkId = this.chainId!;
     }
