+++ src/constant/index.ts
@@ -862,3 +862,14 @@ export const BRAND_ALIAN_TYPE_TEXT = {
   [WALLET_BRAND_TYPES.MATHWALLET]: WALLET_BRAND_CONTENT.MATHWALLET.name,
   [WALLET_BRAND_TYPES.TRUSTWALLET]: WALLET_BRAND_CONTENT.TRUSTWALLET.name,
 };
+
+export const GNOSIS_SUPPORT_CHAINS = [
+  CHAINS_ENUM.ETH,
+  CHAINS_ENUM.BSC,
+  CHAINS_ENUM.POLYGON,
+  CHAINS_ENUM.GNOSIS,
+  CHAINS_ENUM.AVAX,
+  CHAINS_ENUM.OP,
+  CHAINS_ENUM.ARBITRUM,
+  CHAINS_ENUM.AURORA,
+];
