+++ src/ui/models/index.ts
@@ -16,6 +16,7 @@ import { chains } from './chains';
 import { whitelist } from './whitelist';
 import { swap } from './swap';
 import { customRPC } from './customRPC';
+import { safe } from './safe';
 
 export interface RootModel extends Models<RootModel> {
   app: typeof app;
@@ -34,6 +35,7 @@ export interface RootModel extends Models<RootModel> {
   whitelist: typeof whitelist;
   swap: typeof swap;
   customRPC: typeof customRPC;
+  safe: typeof safe;
 }
 
 export const models: RootModel = {
@@ -53,4 +55,5 @@ export const models: RootModel = {
   whitelist,
   swap,
   customRPC,
+  safe,
 };
